Add defensive programming to admin only pages

+++ app.py
@@ -209,25 +209,39 @@ def add_category():
 @app.route("/edit_category/<category_id>", methods=["GET", "POST"])
 @login_required
 def edit_category(category_id):
-    if request.method == "POST":
-        submit = {
-            "category_name": request.form.get("category_name")
-        }
-        mongo.db.categories.replace_one({"_id": ObjectId(category_id)}, submit)
-        flash("Category Successfully Updated")
-        return redirect(url_for("get_categories"))
+    # admin-only page
+    if session["user"] == "admin":
+        # update the category
+        if request.method == "POST":
+            submit = {
+                "category_name": request.form.get("category_name")
+            }
+            mongo.db.categories.replace_one({"_id": ObjectId(category_id)}, submit)
+            flash("Category Successfully Updated")
+            return redirect(url_for("get_categories"))
 
-    category = mongo.db.categories.find_one({"_id": ObjectId(category_id)})
-    return render_template("edit_category.html", category=category)
+        category = mongo.db.categories.find_one({"_id": ObjectId(category_id)})
+        return render_template("edit_category.html", category=category)
+
+    # user is not admin
+    flash("You do not have access to edit categories!")
+    return redirect(url_for("get_houseplants"))
 
 
 # function to delete category from database
 @app.route("/delete_category/<category_id>")
 @login_required
 def delete_category(category_id):
-    mongo.db.categories.delete_one({"_id": ObjectId(category_id)})
-    flash("Category Successfully Deleted")
-    return redirect(url_for("get_categories"))
+    # admin-only page
+    if session["user"] == "admin":
+        # update the category
+        mongo.db.categories.delete_one({"_id": ObjectId(category_id)})
+        flash("Category Successfully Deleted")
+        return redirect(url_for("get_categories"))
+
+    # user is not admin
+    flash("You do not have access to delete categories!")
+    return redirect(url_for("get_houseplants"))
 
 
 if __name__ == "__main__":
